Add relocated `cfg/eval/real/tasks.py` and reroute consumers
Completes the move (d098204 dropped `cfg/tasks.py`): new dependency-free `tasks.py` in the eval/real
package, all consumers rerouted, and `inference` lazy-imported in `cfg.eval.run` to break the
`gui.eval` <-> `cfg.eval` cycle the move introduces.
