+++ positronic/gui/eval.py
@@ -9,7 +9,7 @@ import dearpygui.dearpygui as dpg
 import numpy as np
 
 import pimm
-from positronic.cfg.tasks import SCISSORS_TASK, SPOONS_TASK, TOWELS_TASK, UNIFIED_TASK
+from positronic.cfg.eval.real.tasks import SCISSORS_TASK, SPOONS_TASK, TOWELS_TASK, UNIFIED_TASK
 from positronic.dataset.edits import EditedDataset
 from positronic.dataset.local_dataset import LocalDataset
 from positronic.policy.harness import Directive
